elasticsearch works!

+++ app/Services/Book/DeleteBookService.php
@@ -19,9 +19,9 @@ final readonly class DeleteBookService
     public function execute(int $bookId): void
     {
         $this->transactionManager->transaction(function () use ($bookId): void {
-            $book = $this->bookRepository->findModel($bookId);
+            $book = $this->bookRepository->getById($bookId);
 
-            if ($book->cover_image !== null && $book->cover_image !== '') {
+            if ($book !== null && $book->cover_image !== null && $book->cover_image !== '') {
                 Storage::disk('public')->delete($book->cover_image);
             }
